requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
reportlab==4.2.5
gunicorn==23.0.0
playwright==1.44.0
